requirements.txt
Flask==2.0.1
requests==2.25.1
python-dotenv
groq
numpy